Rename index for easier finding
